Fix multi-language unique script support
This adds better support for languages with a unique script
(such as Korean) when combined with other scripts (such as Latin).

Instead of treating these unique-script languages different from
the rest, they are now combined with other scripts and the top
script is chosen to test languages for. In the case of a script
with only a single language, that language is immediatly returned.

In addition, this change removes Han characters from the Japanese
detection. Previously, Chinese and Japanese would yield very
similar results, now the difference is much bigger.

This fixes a problem described by user lifthrasiir on Hackernews,
and includes two new tests as described there.

Source: https://news.ycombinator.com/item?id=8405672.