Fix results for digit-only and white-space-only input
As the algorithm ignores characters which occur in all or
most languages, `getTopScript` returned `cmn` for these
values (for example, `123 456 789`), as it is the first
script to test and other scripts to reach higher scores.

This bug-fix changes that by checking for a score of `0`, and
returning the `und` string/tuple.

Closes GH-18.